+++ workspaces/azure-devops/plugins/catalog-backend-module-azure-devops-annotator-processor/src/index.ts
@@ -1,0 +1,9 @@
+/**
+ *
+ * @packageDocumentation
+ *
+ * * The Azure DevOps Annotator Processor backend module for the Catalog plugin.
+ */
+
+export { catalogModuleAzureDevopsAnnotatorProcessor as default } from './module';
+export { AzureDevOpsAnnotatorProcessor } from './processor';