+++ workspaces/entity-feedback/plugins/entity-feedback/src/components/FeedbackResponseTable/FeedbackResponseTable.tsx
@@ -21,16 +21,25 @@ import { FeedbackResponse } from '@backstage-community/plugin-entity-feedback-co
 import Chip from '@material-ui/core/Chip';
 import { makeStyles } from '@material-ui/core/styles';
 import CheckIcon from '@material-ui/icons/Check';
+import Typography from '@material-ui/core/Typography';
 import React from 'react';
 import useAsync from 'react-use/esm/useAsync';
 
 import { entityFeedbackApiRef } from '../../api';
+import { Comments } from '../FeedbackResponseDialog';
 
 type ResponseRow = Omit<FeedbackResponse, 'entityRef'>;
 
 const useStyles = makeStyles(theme => ({
   consentCheck: {
     color: theme.palette.status.ok,
+  },
+  listItem: {
+    padding: '0',
+    marginTop: theme.spacing(1),
+  },
+  list: {
+    paddingLeft: '0',
   },
 }));
 
@@ -83,15 +92,53 @@ export const FeedbackResponseTable = (props: FeedbackResponseTableProps) => {
         <>
           {(response.response || '')
             .split(',')
-            .map(v => v.trim()) // removes whitespace
+            .map((v: string) => v.trim()) // removes whitespace
             .filter(Boolean) // removes accidental empty entries
-            .map(res => (
+            .map((res: string) => (
               <Chip key={res} size="small" label={res} />
             ))}
         </>
       ),
     },
-    { title: 'Comments', field: 'comments', width: '40%' },
+    {
+      title: 'Comments',
+      field: 'comments',
+      width: '40%',
+      render: (response: ResponseRow) => {
+        // Check if comment is a stringified object
+        let parsedComment;
+        try {
+          parsedComment =
+            response?.comments && (JSON.parse(response.comments) as Comments);
+        } catch (e) {
+          // If parsing fails, assume it's a regular string
+          parsedComment = response.comments;
+        }
+        return (
+          <div>
+            {typeof parsedComment === 'object' ? (
+              <ul className={classes.list}>
+                {Object.entries<string>(parsedComment.responseComments)?.map(
+                  ([key, value]) => (
+                    <li key={key} className={classes.listItem}>
+                      <strong>{key}:</strong> {value}
+                    </li>
+                  ),
+                )}
+                {parsedComment.additionalComments && (
+                  <li className={classes.listItem}>
+                    <strong>additional:</strong>{' '}
+                    {parsedComment.additionalComments}
+                  </li>
+                )}
+              </ul>
+            ) : (
+              <Typography>{parsedComment}</Typography>
+            )}
+          </div>
+        );
+      },
+    },
   ];
 
   if (error) {