+++ workspaces/entity-feedback/plugins/entity-feedback/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @backstage-community/plugin-entity-feedback
 
+## 0.3.2
+
+### Patch Changes
+
+- 804a245: Fix package metadata
+- Updated dependencies [804a245]
+  - @backstage-community/plugin-entity-feedback-common@0.1.7
+
+## 0.3.1
+
+### Patch Changes
+
+- c2cc0b9: Update dependencies to Backstage v1.28.3
+- Updated dependencies [c2cc0b9]
+  - @backstage-community/plugin-entity-feedback-common@0.1.6
+
+## 0.3.0
+
+### Minor Changes
+
+- f4e4277: Add comment boxes for each checked option in the Feedback Dialog box. Display those option responses as bullet points in the Feedback Response table.
+
 ## 0.2.19
 
 ### Patch Changes