+++ workspaces/vault/plugins/vault/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-vault
 
+## 0.1.31
+
+### Patch Changes
+
+- 6757373: Backstage v1.28.4 version bump. Also addressed some of the upstream deprecations
+
 ## 0.1.30
 
 ### Patch Changes